Add Kubernetes config for deploying PostgreSQL
This commit adds configuration for deploying a PostgreSQL container
into a Kubernetes cluster. It also contains a LoadBalancer service to
expose the container externally for testing and verification purposes,
as well as a PersistentVolume to store the contents of the database.

This database will be used by the Kubernetes example BuildGrid
deployment.