Add Kubernetes config for deploying BuildGrid
This config actually deploys three separate services; `bgd-execution`
for Execution/Operations/Bots services, `bgd-cas` for CAS and
ByteStream, and `bgd-cache` for the ActionCache.

It is possible to connect a local worker to `bgd-execution` and
`bgd-cas` to run an experimental build using this BuildGrid.

The configuration is pretty much the simplest possible config with
this split of services. The scheduler uses the PostgreSQL service
deployed by the config added in !1. The ActionCache is in-memory, and
CAS uses a 30GB volume for disk storage.