Add an Nginx proxy for handling incoming requests to BuildGrid
This proxy is used to provide a single endpoint for the whole remote
execution grid, which makes it easier to use with tools like Bazel which
impose restrictions on where services are located.

The proxy handles routing requests to the correct services inside the
Kubernetes deployment.

This commit also stops exposing the internal services externally.