Add a Redis server to the Kubernetes example
This commit adds a Redis server to the example Kubernetes deployment.
This is used to provide a persistent-ish backend for the ActionCache
service, which also allows the ActionCache to be scaled if so desired.

Restarting the Redis container will empty the cache, so it isn't fully
persistent. It simply serves as an example of a restartable and scalable
ActionCache.